Export projets route
